Rename the terminfo file and add comments explaining the intended direction: this terminal is going to try to be compatible with whatever xterm terminfo would look like so that it can be used interchangeable with other xterm-compatible terminal emulators like iTerm2 and hterm.
The terminfo may still be useful to bootstrap a system without a modern
xterm terminfo. Currently, it is based on the terminfo for st (the
original project) with 256-color support forcibly merged. The goal will
be to adjust this to be an increasingly precise match for whatever other
xterm-compatible terminals are doing.