Clean up the terminfo record we have to start more assiduously emulating xterm's behavior in all cases for broader compatibility with other xterm-emulating terminals (iTerm2 and hterm for example).
First, I've moved the terminfo to a filename that makes it more clear
what the intent is. I've also enhanced the comments at the top to talk
about how to use this file as a record of the intended behavior of mt.

Second, I've sorted and removed duplicate entries for all the
properties, and merged the meta components.

Third, I've systematically compared the values with an old but not
ancient xterm set of values. Where there was an easy merge, I've just
annotated the differences. Where there are conflicting values I've
marked the conflict.

This will hopefully serve as a documentation basis for slowly
harmonizing the behavior of mt in a few of the corner cases with that of
xterm (where that makes sense). I still need to go through a much more
modern copy of xterm's info to make sure nothing important is
missing/changed and I'm also going to look at what specific, popular
xterm-emulating terminals do in some of these cases to help guide the
eventual behavior here.