Partially revert e50b794 and introduce a better fix. Rather than completely ignoring the mode changes, leave that information available but change the keybindings for both the keypad delete and the normal delete keys from X to trigger the same escape code sequence. After testing xterm's behavior, it does not appear to produce different escape codes based on modes for delete.
I'm working through auditing other common keys to understand better what
applications are likely to expect in practice and what the most
compatible and effective escape codes are to send.

One thing to be aware of is that I have no real interest in supporting
unusual terminal emulation modes where the keypad is disabled and
enabled differently in these modes unless someone lets me know *why*
this would be useful. I don't realy use a keypad at all. So if you do
use a keypad and are interested in carefully testing this behavior,
don't hesitate to send pull requests. But keep in mind that maximum
compatibility with hardware or terminals isn't the primary goal --
compatibility with common, modern hardware and common, modern software
is. For example, zsh, bash, readline, and irssi need to work out of the
box, regardless of how buggy they are.