Make backspace match xterm and remove a weird special case for it.
The old version of this seemed to work in all the applications I use but
it definitely doesn't match the terminfo where this does. This also
tests fine for me. With this I get to remove another discrepancy from
the record of terminfo codes.

One thing to note is that I don't know why mod1+backspace was escaping
the delete. I've deleted that as I don't see anything that really
resembles this with xterm.

However, I can't explain xterm's behavior when using mod1+backspace so
I haven't tried to emulate it and instead backspace just always sends
^H to the terminal which seems like a pretty good state.