Configure shell in a more obvious way.
Provide a single configuration option: higher priority than system
config, lower than -c. This is similar to the old utmp, but without the
weird name and the sucklessism that utmp never became $SHELL (because it
is a helper program).

Hardcode the /bin/sh fallback when detecting preferred shell fails.

Spend a few more lines to make the fallback sequence (IMO) less cryptic.