Remove tabspaces config flag, in favor of dynamic tabstop management.
The initial tabstops state is described by terminfo, so this setting
could not be effectively used without modifying terminfo.

Programs can set the terminal tabstops, e.g. the POSIX `tabs`
command, which is the typical way to adjust this behavior.

Configuring `tabspaces` have one advantage: because it's an interval rather
than a list of columns, it's clear where to insert new tabstops if the
terminal is widened.

I simulate this by inferring the width from the first tabstop, which
will do the right thing if tabs have uniform width.

Various terminals' behavior when widening after `tabs 4`:

  - mt previously added new tabstops 8 apart
  - urxvt resets all tabs to 8 wide
  - xterm+gnome-terminal make tabs past the old width wrap onto a new line