Mostly revert d7ff9721: backspace sends DEL again.
The previous commit made backspace send ^H, for consistency with xterm's
terminfo. Unfortunately this broke backspace on my laptop.

It turns out xterm's terminfo itself is not consistent across systems:
`infocmp xterm | grep kbs` reports \177 on my ubuntu machines.

This matches debian's policy:

    https://www.debian.org/doc/debian-policy/ch-opersys.html#s9.8

which redhat also followed (10 years ago, but I bet it's still the same)

    https://bugzilla.redhat.com/show_bug.cgi?id=142659

On the other hand, some distros (e.g. Arch) and non-linuxes use ^H.
The upstream sources have a pretty sad comment on this:

    http://invisible-island.net/xterm/terminfo-contents.html#tic-xterm_kbs

I can't explain why it works on some systems even with the "wrong"
string, but maybe we can lean on that for now?

I think the correct fix is for mt to read the terminfo database and do
what it says, left a TODO for now.