Extract ANSI escape sequence parser into a separate class.
The new implementation is based on the VT100 state machine described at:
https://vt100.net/emu/dec_ansi_parser, with modifications for extensions
that mt already supports, such as ESC-K to set the title.
(*Most* future features shouldn't require changes, as they're e.g. CSI
sequences using standard syntax.)

This is a big patch as the parser was previously intertwined with much
of the other code. I've "modernized" many of the affected callsites, but
tried not to be too invasive.

The new parser is unit-testable but not actually unit-tested yet - happy
to write these but would like some guidance on how you'd like these to
look.
I verified the behavior by comparing vttest output for mt with and
without the patch. They're identical apart for the DECSCA test (11.1.2.4)
where the new behavior is more correct.

Some minor changes made for implementation convenience:

- When reading sixel graphics (which are ignored), we now utf-8-decode the
inputs before processing them (and with MODE_PRINT, write utf-8 output).
Valid sixels are printable ASCII, this only affects malformed streams.
- With MODE_PRINT, we no longer replace non-printable non-control
characters with UTF_INVALID in the output file.
- There's no longer limits on string lengths or number of parameters to
escape sequences.